Revert commits that cause error if no repo is found
If the repo is missing then it should automatically be recreated when it's
next run. The reason this was probably not happening before is due to file
permissions differences between buildpack and legacy scraper runs. This
permission issue is now fixed so this shouldn't be a problem anymore.

Revert commits:
8a593f0f418da05623ac30c0c2382d0fdb9d22e2
a72b445425994c7bb0f34e7a30db09c822cfde0e
8f0999cbe461ad18db6e1a2133a0687c23cb5725