add script + styles to make table fade effect match scroll position
The script calculates whether the table needs a fade effect and where
it should be placed. It does this on load, and then recalculates as
the box is scrolled.

The fade effect style is added as classes are toggled on/off based
on the calculation.

This makes the fade effect more helpful, as it changes to suit

There is an outstanding bug: When clicking the .nav-tabs to
move  between panels, the fade-right class is always added because
the calculation happens too early, before the tab is visible.