fix widths of the scraper info elements so they're easy to scan
This keeps all the icons and the scraper language info inline so
that it can all be scanned easily. For easy scraper the info is in
a horizontal line, and between the scrapers stacked on top of one
another, each equivalent icon is in a vertical line.

see https://github.com/openaustralia/morph/pull/651#issuecomment-96889536