keep the icons in the same place when scaper language not defined
This catches a corner case where the scraper has no defined
language so the icons moved into the place of the language element
and the vertical alignment was broken.

This looks very strange on very small screens with the missing
language element, so I've only made it kick in once screens are
a little wider.