display scraper repo/file in "rep / file" format
This simplifies the layout, and presents the file in a familiar
way for developers. This display is less explanitory, but we don't
think that will be a big problem for scraper author.

This removes the headings from the section and merges the separate
repo and file links into a single 'repo_name / filename.rb' line,
still two separate links.