Add example scraped domains to scraper lists (static prototype)
Know what domains are being scraped is an important and useful
piece of information for understanding what a scraper does.

Showing this information in the scraper lists should help citizens
find the scrapers they are looking for.

This commit adds a static text version of this for use in designing
this interface.

relates to #677