Move the scraper 'errored' label to right-hand side
In scraper lists, move the 'errored' label from the left hand side,
where it dumps in the label, to the right-hand side.

This keep the heading more prominent that the label and makes the
scraper list layout more less variable. This groups all the similar
scraper status information together.

relates to #689