Show the 'errored' label where the 'running' icon sits
In scraper lists, make the scraper status information layout
simpler by vertically aligns the 'running' or 'errored' labels.
This should make the scraper information on the right-hand side
easier to read.

If the scraper doesn't have the running indicator, don't
insert it's markup wrapper. This means that if the errored label
is showing, it sits in the same vertical alignment as the running
label.

fixes #689