Make the download notice layout css more robust
This is so that the layout of the section is consistent whether
or not the notice is present.

This moves the margin setting to downloads block off the nav tabs
from the alert, and onto the wrapping block div. Then the internal
and bottom spacing is added on the inner elements.