Only show users that have a 'name'
We'd really like to only show users here who have a custom image
instead of the default github identicon. But we can't work out a
way to achieve that without nasty hacks. We've seen that pretty
consistently users who have a 'name' don't have the identicon. So
for now, we'll use 'name' as a proxy for 'avaratar is not identicon'.

Remove the TODO for this change.

See discussion at
https://github.com/openaustralia/morph/pull/768