Don't show featured supporters on landing page when logged out
If you're logged out on the landing page you're more than likely
not yet a user of morph.io. So, getting a social incentive to become
a supporter doesn't make sense yet. You should start using morph.io first,
get to like it, and then you're likely to become a supporter with a little
nudge.