Front load the important bit of the sentence
Putting the key element that people who should read this sentence
need to see at the front makes it easier to scan. This might make
the sentence slightly verbose, but it makes it more likely to be
read. There’s lots of tips like this in the GDS’s guides
https://www.gov.uk/guidance/content-design/writing-for-gov-uk