Slow the user down as they read the danger options
Use the layout to encourage the user to read the descriptions of
the dangerous options, so they better understand what's going to
happen once they act. Making the wrong choice here could mean
the user looses data they’ve scraped irretrievably, so we should
help them slow down and consider.

This commit moves the buttons to the right hand side of the box
on wide screens, away from the text descriptions, so the user’s
eye has to jump back across to read and make sure they're picking
the correct option.