Increase allowed sidekiq memory usage to 4GB before monit restarts it
This is to stop it being killed so readily, which often results in
duplicate jobs being created - #926.

Not having it restarted by monit means that the oomkiller is invoked
which has its own problems. Namely that scrapers are seemingly killed
which results in them appearing stuck running in the interface - #910.

The reason the memory is blowing out in the first place appears to be
because of scrapers outputting heaps of log lines - #919.