Visually distinguish dangerous setting buttons with colour
Make the ‘clear database’ setting orange, as distinct from the 'delete
scraper' button’s red.

These buttons are still too similar in appearance. People who understand
the settings, because they are familiar with them for example, can
easily make a slip by selecting the wrong option. You're probably using
this setting as part of some other task, and are likely to be moving
quick quickly.

Don Norman calls these kinds of errors “Description-similarity slip” in
The Design of Everyday Things:
http://www.jnd.org/books/design-of-everyday-things-revised.html

This commit using the bootstrap helper class to change the colour of the
button to a 'warning orange' from a 'danger red'. This also helps people
understand that the 'delete scraper' option is indeed a bit more severe.