Don't start a run if it's already been marked as finished
If the run is marked as finished it means it got a very long way through
the run and is probably retrying. The tasks that happen after a run is
marked as finished are just refreshing scraper data.

However if we instead rerun a failed run that got all this way then it
becomes very confusing and we show really weird stuff to the user, e.g.
the scraper says it's finished but the console will be outputting stuff
(because the run is actually running again).