Force delete of container on clean up
This is a bit of a stab in the dark at fixing #1034.

The error in #1034 is similar to:

		Docker::Error::NotFoundError: open /var/lib/docker/containers/e0f025881bfdfb5bb5334b4da26c535bad8aebf1ef444e79db1ee80ee438cc46/e0f025881bfdfb5bb5334b4da26c535bad8aebf1ef444e79db1ee80ee438cc46-json.log: no such file or directory

But I've observed that before that happens an error similar to this
occurs:

		Docker::Error::ServerError: Driver aufs failed to remove root filesystem ab53f1e061b9eb03284f35919d1e0a7c7edf4b2c8b1ce12a6c0afec55f724608: rename /var/lib/docker/aufs/mnt/b133538fc8ba5d54e81f67b435a9cf4fc6cd84e42f860baaba864d02d90bebf8 /var/lib/docker/aufs/mnt/b133538fc8ba5d54e81f67b435a9cf4fc6cd84e42f860baaba864d02d90bebf8-removing: device or resource busy

The `Docker::Error::NotFoundError` happens on a *retry* after the above
error, which happens on this line, while trying to clean up the
container.

I want to see if this will allow the delete to happen without
complaining about not being able to remove the filesystem. Hopefully the
filesystem is then tidied up in our usual Docker clean up.