Add rake task to remove images in "created" state
As part of #1104 we've discovered hundreds of images that are sitting in
the "created" state. These all seem to be failed container compile steps
and can be safely deleted.

So we're adding this rake task to see if cleaning them up also reclaims
disk space. If it does then we might consider making this part of the
regular docker clean up. If it doesn't then we can remove this task.

Of course even better would be to fix the underlying problem so that
these failed container compiles don't leave around unused Docker
containers :smile: