Add rake task to delete old stopped morph docker containers
As part of #1104 we've found lots of old stopped containers. Most of
these are from failed container compile steps so seem safe to delete.

This task only deletes images older than 1 week because sometimes morph
problems mean that stopped images hang around and need to be properly
cleaned up.

If this task results in a significant amount of disk space being freed
then we should consider making it part of the regular docker clean up.
If it doesn't then this should probably be removed.