Change stopped container clean up task to be more broad
About half of the stopped containers I wanted to clean up just now were
not being seen by this task. I'm not sure why they're not ancestors of
the morph base image so this change is necessary to pick them up.

Weirdly they're definitely from failed buildpack compilation because
they've got the same command as the rest of the containers that were
picked up.

It's also not useful to just clean up old containers. If they're
orphaned they'll never be cleaned up so why not do it now. By checking
there's no associated run it means we're not getting scrapers that are
just waiting to have their logs streamed off and cleaned up.