Hardcode path to sslmate files
We don't have certificates for dev.morph.io - even adding an exception
this will break faye on the test VM.

This is probably something we want to fix but it will likely mean making
the local VM pretend it's actually running at morph.io instead. The
other option is expanding the certificate to cover that domain but with
sslmate each domain costs us more so it's probably not worth it.