Disable Clear database button if scraper is empty
This commit aims to fix the issue https://github.com/openaustralia/morph/issues/983
If the scraper is empty, the button to clear the database on the settings page must be disabled for ease to the user. 
So, in this commit, the design and action of the button has been changed to make it disabled if the scraper is empty.