When creating a Run also create associated Metric
We always want a run to have an associated metric instance, so
automatically create one whenever a run is created.

This was triggered by a problem destroying runs because they had
multiple metrics associated with them. This doesn't actually fix that
problem, but it takes a step towards it by centralising the creation of
metrics.