Don't create metrics in Morph::Runner
This was causing issues with runs having multiple metrics associated
with it, which is supposed to be a one-to-one relationship. The metric
for a run is now created when the run is created, so we can just update
it here.

Ideally this fix would have a regression test, but as things currently
stand this method has quite a few hard to test side-effects. I had a bit
of a go at pulling it apart but decided in this case done is better than
perfect.