Much simplified redis install
The default configuration of redis on ubuntu 16.04 now matches exactly
what we were using in our configuration. We're going to a slightly older
version of redis but one that is still a lot newer than when morph
was first made. I figure it's better to stick with the defaults if at all
possible.