Inline RunAbility
This was a slightly strange way to use the cancan Ability so have inlined the code instead which makes it clearer what's going on anyway.

There's now a slightly mixed concern between the check for the read-only mode inside Ability classes and outside but I'm not too worried about that.