Turn on auto_paginate for all requests through octokit_client
Because we weren't doing this everywhere it was causing scrapers to fail
because it thought it didn't have access when in fact
it did. It's just that the list of repositories that it had
access to was too large so wasn't getting returned in
a single github lookup. Hence the need for auto_paginate.