Deprecate use of organisation api keys
With private repos we can use user's permissions on a repo to easily determine whether they should have read access to the repo. This relationship is a lot more muddy with organisations.

Organisations can own lots of repos but who should have access to those API keys? It's all a bit complicated and frankly not worth the effort. So, it's easier to deprecate organisation API keys.

There are only a very small handful of organisation API keys in use at the moment. So, it's not a big problem really.